Optimized out the allocation
Encoders are now the disposable state struct for each Handle call.  A sync pool of encoders replaces the sync pool of buffers.

This allows the encoder to store additional state per Handle() call, including a slice of currently open groups, which is needed to call ReplaceAttr.

Benchmarking shows 0 allocations, regardless of whether ReplaceAttr is set or not.  Setting a ReplaceAttr does add some overhead, similar to the slog.TextHandler.

Benchmarking also shows a slight performance improvement over the original code (using the buf pool), when ReplaceAttrs is not set.